# Implement build cmd in rust 