--- a/src/meta_memcache/extras/client_wrapper.py
+++ b/src/meta_memcache/extras/client_wrapper.py
@@ -1,4 +1,4 @@
-from typing import Any, Dict, List, Optional, Set
+from typing import Any, Dict, List, Optional
 
 from meta_memcache.commands.high_level_commands import HighLevelCommandsMixin
 from meta_memcache.configuration import ServerAddress
@@ -6,11 +6,9 @@ from meta_memcache.connection.pool import PoolCounters
 from meta_memcache.interfaces.router import FailureHandling, DEFAULT_FAILURE_HANDLING
 from meta_memcache.interfaces.cache_api import CacheApi
 from meta_memcache.protocol import (
-    Flag,
-    IntFlag,
     Key,
     ReadResponse,
-    TokenFlag,
+    RequestFlags,
     WriteResponse,
 )
 
@@ -33,32 +31,24 @@ class ClientWrapper(HighLevelCommandsMixin):
     def meta_multiget(
         self,
         keys: List[Key],
-        flags: Optional[Set[Flag]] = None,
-        int_flags: Optional[Dict[IntFlag, int]] = None,
-        token_flags: Optional[Dict[TokenFlag, bytes]] = None,
+        flags: Optional[RequestFlags] = None,
         failure_handling: FailureHandling = DEFAULT_FAILURE_HANDLING,
     ) -> Dict[Key, ReadResponse]:
         return self.client.meta_multiget(
             keys=keys,
             flags=flags,
-            int_flags=int_flags,
-            token_flags=token_flags,
             failure_handling=failure_handling,
         )
 
     def meta_get(
         self,
         key: Key,
-        flags: Optional[Set[Flag]] = None,
-        int_flags: Optional[Dict[IntFlag, int]] = None,
-        token_flags: Optional[Dict[TokenFlag, bytes]] = None,
+        flags: Optional[RequestFlags] = None,
         failure_handling: FailureHandling = DEFAULT_FAILURE_HANDLING,
     ) -> ReadResponse:
         return self.client.meta_get(
             key=key,
             flags=flags,
-            int_flags=int_flags,
-            token_flags=token_flags,
             failure_handling=failure_handling,
         )
 
@@ -67,9 +57,7 @@ class ClientWrapper(HighLevelCommandsMixin):
         key: Key,
         value: Any,
         ttl: int,
-        flags: Optional[Set[Flag]] = None,
-        int_flags: Optional[Dict[IntFlag, int]] = None,
-        token_flags: Optional[Dict[TokenFlag, bytes]] = None,
+        flags: Optional[RequestFlags] = None,
         failure_handling: FailureHandling = DEFAULT_FAILURE_HANDLING,
     ) -> WriteResponse:
         return self.client.meta_set(
@@ -77,40 +65,30 @@ class ClientWrapper(HighLevelCommandsMixin):
             value=value,
             ttl=ttl,
             flags=flags,
-            int_flags=int_flags,
-            token_flags=token_flags,
             failure_handling=failure_handling,
         )
 
     def meta_delete(
         self,
         key: Key,
-        flags: Optional[Set[Flag]] = None,
-        int_flags: Optional[Dict[IntFlag, int]] = None,
-        token_flags: Optional[Dict[TokenFlag, bytes]] = None,
+        flags: Optional[RequestFlags] = None,
         failure_handling: FailureHandling = DEFAULT_FAILURE_HANDLING,
     ) -> WriteResponse:
         return self.client.meta_delete(
             key=key,
             flags=flags,
-            int_flags=int_flags,
-            token_flags=token_flags,
             failure_handling=failure_handling,
         )
 
     def meta_arithmetic(
         self,
         key: Key,
-        flags: Optional[Set[Flag]] = None,
-        int_flags: Optional[Dict[IntFlag, int]] = None,
-        token_flags: Optional[Dict[TokenFlag, bytes]] = None,
+        flags: Optional[RequestFlags] = None,
         failure_handling: FailureHandling = DEFAULT_FAILURE_HANDLING,
     ) -> WriteResponse:
         return self.client.meta_arithmetic(
             key=key,
             flags=flags,
-            int_flags=int_flags,
-            token_flags=token_flags,
             failure_handling=failure_handling,
         )