--- a/src/meta_memcache/settings.py
+++ b/src/meta_memcache/settings.py
@@ -5,4 +5,7 @@ DEFAULT_MARK_DOWN_PERIOD_S = 5.0
 
 DEFAULT_READ_BUFFER_SIZE = 4096
 
-MAX_KEY_SIZE = 250
+# Max key is 250, but when using binary keys will be b64 encoded
+# so take more space. Keys longer than this will be hashed, so
+# it's not a problem.
+MAX_KEY_SIZE = 187