Build cmd in rust

## Motivation / Description
The second cpu-intensive part of the request processing is building
the cmd.

Also instead of building dicts of flags we can use a single
flags object, which also simplifies the API of the lower
commands.

I chose to still built a single flags object, but we could
explore building one flags object per meta-command, as the
flags that they support differ, and it could lead to a more
type-safe low-level implementation.

## Performance:
* Initial (before all this patchsets):
multithreaded: Overall: 110779.55 RPS / 9.03 us/req
singlethreaded: Overall: 111545.63 RPS / 8.96 us/req

* Rust only for response parsing
multithreaded: Overall: 245898.34 RPS / 4.07 us/req
singlethreaded: Overall: 246165.19 RPS / 4.06 us/req

* Now (rust also for build_cmd)
multithreaded: Overall: 319587.03 RPS / 3.13 us/req
singlethreaded: Overall: 323101.77 RPS / 3.10 us/req

## Speedscope:
* Before: 1.18us
![Screenshot 2023-11-07 at 23 47 53](https://github.com/RevenueCat/meta-memcache-py/assets/1917/f98a56a1-d143-4bf4-813e-164eee2ec435)

* After: 456ns
![Screenshot 2023-11-07 at 23 45 56](https://github.com/RevenueCat/meta-memcache-py/assets/1917/1861b30c-6432-4815-9895-9e0a54544e46)